exercice-4/level-4/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes -o requirements.txt
absl-py==2.3.1
    # via
    #   keras
    #   tensorboard
    #   tensorflow
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
astunparse==1.6.3
    # via tensorflow
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
charset-normalizer==3.4.3
    # via requests
click==8.2.1
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   uvicorn
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.116.1
    # via level-4
fastapi-cli==0.0.11
    # via fastapi
fastapi-cloud-cli==0.1.5
    # via fastapi-cli
flatbuffers==25.2.10
    # via tensorflow
gast==0.6.0
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
grpcio==1.74.0
    # via
    #   tensorboard
    #   tensorflow
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h5py==3.14.0
    # via
    #   keras
    #   tensorflow
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
jinja2==3.1.6
    # via fastapi
keras==3.11.3
    # via
    #   level-4
    #   tensorflow
libclang==18.1.1
    # via tensorflow
markdown==3.9
    # via tensorboard
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.3
    # via
    #   keras
    #   tensorflow
namex==0.1.0
    # via keras
numpy==2.3.3
    # via
    #   h5py
    #   keras
    #   ml-dtypes
    #   pandas
    #   tensorboard
    #   tensorflow
opt-einsum==3.4.0
    # via tensorflow
optree==0.17.0
    # via keras
packaging==25.0
    # via
    #   keras
    #   tensorboard
    #   tensorflow
pandas==2.3.2
    # via level-4
pillow==11.3.0
    # via tensorboard
protobuf==6.32.1
    # via
    #   tensorboard
    #   tensorflow
pydantic==2.11.9
    # via
    #   fastapi
    #   fastapi-cloud-cli
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via uvicorn
python-multipart==0.0.20
    # via fastapi
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via uvicorn
requests==2.32.5
    # via tensorflow
rich==14.1.0
    # via
    #   keras
    #   rich-toolkit
    #   typer
rich-toolkit==0.15.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.6.4
    # via fastapi-cloud-cli
sentry-sdk==2.38.0
    # via fastapi-cloud-cli
setuptools==80.9.0
    # via
    #   tensorboard
    #   tensorflow
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow
sniffio==1.3.1
    # via anyio
starlette==0.47.3
    # via fastapi
tensorboard==2.20.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.20.0
    # via level-4
termcolor==3.1.0
    # via tensorflow
typer==0.17.4
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   optree
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   starlette
    #   tensorflow
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.35.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
uvloop==0.21.0 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
werkzeug==3.1.3
    # via tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.3
    # via tensorflow
